apps/skill-downloader/pyproject.toml
[project]
name = "skill-downloader"
version = "0.1.0"
description = "Download skill.md files from GitHub repositories"
requires-python = ">=3.12"
dependencies = [
    "gh",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
skill-downloader = "skill_downloader.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skill_downloader"]

[tool.uv.sources]
gh = { workspace = true }
